fix: SentryInternalClient crashing on some log types.
[Log messages from Django](https://docs.djangoproject.com/en/2.0/topics/logging/#django-request)
contain a `request` object which is not JSON serializable. The regular
Raven `Client` has a pretty bulletproof JSON encoder which handles these
objects, but SentryInternalClient didn't use it, so there is an
exception thrown further down the stack in `insert_data_to_database`
when it comes across these objects. This fix forces the Raven JSON
encoding of the event.

This is not a very efficient solution, but it is at least simple.

 Untracked files: